+++ frontend/util/peers.js
@@ -1,8 +1,14 @@
 export function servers(peers_json, system_status) {
-    let servers = [{
-        id: system_status.id,
-        hostname: system_status.hostname,
-    }].concat(peers_json)
+    let servers = [];
+    if(peers_json) {
+        servers = servers.concat(peers_json)
+    }
+    if(system_status) {
+        servers.push({
+            id: system_status.id,
+            hostname: system_status.hostname,
+        });
+    }
     servers.sort((a, b) => a.hostname.localeCompare(b.hostname))
     return servers
 }